🎨 Improve user experience with using AccountsSnapshots derive macro
