+++ examples/fuzz-tests/arbitrary-limit-inputs-5/programs/arbitrary-limit-inputs-5/Cargo.toml
@@ -15,8 +15,10 @@ no-entrypoint = []
 no-idl = []
 no-log-ix-name = []
 idl-build = ["anchor-lang/idl-build", "anchor-spl/idl-build"]
+trident-fuzzing = ["dep:trident-fuzz"]
 
 [dependencies]
+trident-derive-accounts-snapshots = { path = "../../../../../crates/fuzz/derive/accounts_snapshots" }
+trident-fuzz = { path = "../../../../../crates/fuzz", optional = true }
 anchor-lang = "0.30.1"
 anchor-spl = "0.30.1"
-trident-client = { path = "../../../../../crates/client" }
