+++ examples/fuzz-tests/arbitrary-limit-inputs-5/trident-tests/fuzz_tests/Cargo.toml
@@ -19,3 +19,4 @@ path = "../../../../../crates/client"
 
 [dependencies.arbitrary-limit-inputs-5]
 path = "../../programs/arbitrary-limit-inputs-5"
+features = ["trident-fuzzing"]
